Dataset linked to Telecom churn rate
This was made available in a Data Analyst bootcamp as a dataset to be used in a Business Case exercise. The goal was to develope an analysis to help a company to reduce thir churn rate which represents clients that left the company by quiting or not renewing their service contracts.